requirements.txt
discord.py-self
asyncio
python-dotenv